Loner AI
Made a simple AI where the loners will stay in a spot until someone comes too close. The structure of the scripts and stuff are subject to change. I need to research inheritance more before I can condense the code. Until then the format is a little redundant since there is repitition in code